renaming packages and adding new project module
