+++ build.gradle
@@ -2,8 +2,8 @@ plugins {
     id 'java'
 }
 
-group = 'works.hop'
-version = '0.1.0'
+group = 'works.hop.presso'
+version = '1.0-SNAPSHOT'
 
 repositories {
     mavenCentral()
