+++ presso-extn/src/main/java/works/hop/presso/extn/Main.java
@@ -1,0 +1,7 @@
+package works.hop.presso.extn;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
