+++ presso-json/build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'java'
 }
 
-group = 'works.hop.json'
+group = 'works.hop.presso'
 version = '1.0-SNAPSHOT'
 
 repositories {
