--- presso-json/src/main/java/works/hop/json/example/Address.java
@@ -1,16 +1,0 @@
-package works.hop.json.example;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Address {
-
-    String streetAddress;
-    String city;
-    String state;
-    String postalCode;
-}
