--- presso-json/src/main/java/works/hop/json/example/ContactsList.java
@@ -1,6 +1,0 @@
-package works.hop.json.example;
-
-import java.util.LinkedList;
-
-public class ContactsList extends LinkedList<Phone> {
-}
