--- presso-json/src/main/java/works/hop/json/example/ContactsMap.java
@@ -1,6 +1,0 @@
-package works.hop.json.example;
-
-import java.util.LinkedHashMap;
-
-public class ContactsMap extends LinkedHashMap<String, Phone> {
-}
