--- presso-json/src/main/java/works/hop/json/example/EmployeeAr.java
@@ -1,18 +1,0 @@
-package works.hop.json.example;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class EmployeeAr {
-
-    String firstName;
-    String lastName;
-    int age;
-    String emailAddress;
-    Address address;
-    Phone[] contactInfo;
-}
