--- presso-json/src/main/java/works/hop/json/example/PhoneType.java
@@ -1,6 +1,0 @@
-package works.hop.json.example;
-
-public enum PhoneType {
-
-    CELL, HOME, OFFICE, FAX
-}
