+++ presso-json/src/main/java/works/hop/presso/json/JsonLoader.java
@@ -1,0 +1,23 @@
+package works.hop.presso.json;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Objects;
+import java.util.function.Function;
+
+public interface JsonLoader<T> extends Function<InputStream, T> {
+
+    default T classpathResource(String fileName) {
+        try (InputStream is = Objects.requireNonNull(
+                JNodeParser.class.getClassLoader().getResourceAsStream(fileName))) {
+            return apply(is);
+        } catch (IOException e) {
+            try (InputStream is = Objects.requireNonNull(
+                    JNodeParser.class).getResourceAsStream(fileName)) {
+                return apply(is);
+            } catch (IOException e2) {
+                throw new RuntimeException(e2);
+            }
+        }
+    }
+}
