+++ presso-json/src/main/java/works/hop/presso/json/example/ContactsList.java
@@ -1,0 +1,6 @@
+package works.hop.presso.json.example;
+
+import java.util.LinkedList;
+
+public class ContactsList extends LinkedList<Phone> {
+}
