+++ presso-json/src/main/java/works/hop/presso/json/example/ContactsMap.java
@@ -1,0 +1,6 @@
+package works.hop.presso.json.example;
+
+import java.util.LinkedHashMap;
+
+public class ContactsMap extends LinkedHashMap<String, Phone> {
+}
