+++ presso-json/src/main/java/works/hop/presso/json/example/EmployeeKv.java
@@ -1,0 +1,20 @@
+package works.hop.presso.json.example;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.Map;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class EmployeeKv {
+
+    String firstName;
+    String lastName;
+    int age;
+    String emailAddress;
+    Address address;
+    Map<String, Phone> contactInfo;
+}
