+++ presso-json/src/main/java/works/hop/presso/json/example/PhoneType.java
@@ -1,0 +1,6 @@
+package works.hop.presso.json.example;
+
+public enum PhoneType {
+
+    CELL, HOME, OFFICE, FAX
+}
