+++ presso-json/src/main/java/works/hop/presso/json/naming/Entry.java
@@ -1,0 +1,4 @@
+package works.hop.presso.json.naming;
+
+public record Entry<T>(int index, T value) {
+}
