+++ presso-json/src/main/java/works/hop/presso/json/naming/NamingStrategy.java
@@ -1,0 +1,8 @@
+package works.hop.presso.json.naming;
+
+public interface NamingStrategy {
+
+    String resolve(String name);
+
+    String inverse(String name);
+}
