+++ presso-json/src/test/resources/address-to-address2.properties
@@ -1,0 +1,2 @@
+street_address=streetAddress
+zip_code=postalCode
