+++ presso-json/src/test/resources/address2-to-address.properties
@@ -1,0 +1,2 @@
+streetAddress=street_address
+postalCode=zip_code
