+++ presso-json/src/test/resources/contacts-config-list.properties
@@ -1,1 +1,1 @@
-$=[works.hop.json.example.Phone]
+$=[works.hop.presso.json.example.Phone]
