+++ presso-json/src/test/resources/contacts-config-map.properties
@@ -1,1 +1,1 @@
-$=java.util.HashMap<java.util.String, works.hop.json.example.Phone>
+$=java.util.HashMap<java.util.String, works.hop.presso.json.example.Phone>
