+++ presso-json/src/test/resources/employee-config.properties
@@ -1,1 +1,1 @@
-$.contactInfo=[works.hop.json.example.Phone]
+$.contactInfo=[works.hop.presso.json.example.Phone]
