+++ presso-json/src/test/resources/employee_array_list-config.properties
@@ -1,1 +1,1 @@
-$.contactInfo=java.util.ArrayList[works.hop.json.example.Phone]
+$.contactInfo=java.util.ArrayList[works.hop.presso.json.example.Phone]
