+++ presso-json/src/test/resources/employee_key_values-config.properties
@@ -1,1 +1,1 @@
-$.contactInfo=java.util.HashMap<java.lang.String, works.hop.json.example.Phone>
+$.contactInfo=java.util.HashMap<java.lang.String, works.hop.presso.json.example.Phone>
