+++ settings.gradle
@@ -2,6 +2,6 @@ rootProject.name = 'espresso'
 include 'presso-api'
 include 'presso-jetty'
 include 'presso-demo'
-include 'presso-cli'
 include 'presso-json'
+include 'presso-extn'
